Move logged cluster types under pkg/logged-cluster so it's easier to find them
