+++ pkg/logged-cluster/vintagewc/cluster.go
